Added new mode that bridge best of both world of a a BufferedWriter
If the full content of a segment can be hold in memory, we skip the temporary file writing altogether and write directly to destination, eliminating all temporary I/O aspects.

If it does not fit in memory, much less write will be made as lot will be accumulated in the buffer before writing back, optimizing a lot of time lost in doing I/O operations.